jwt 토큰 적용

+++ src/main/java/com/sparta/myblogbackend/controller/BlogController.java
@@ -4,6 +4,8 @@ import com.sparta.myblogbackend.dto.BlogRequestDto;
 import com.sparta.myblogbackend.dto.BlogResponseDto;
 import com.sparta.myblogbackend.jwt.JwtUtil;
 import com.sparta.myblogbackend.service.BlogService;
+import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,8 +19,12 @@ import java.util.List;
 public class BlogController {
     private BlogService blogService;
 
+    private final JwtUtil jwtUtil;
+
     @PostMapping("/blog")
-    public BlogResponseDto createBlog(@RequestBody BlogRequestDto requestDto) {
+    public BlogResponseDto createBlog(@RequestBody BlogRequestDto requestDto, HttpServletResponse res) {
+        String token = jwtUtil.createToken(requestDto.getUsername());
+        jwtUtil.addJwtToCookie(token, res);
         return blogService.createBlog(requestDto);
     }
 
@@ -34,12 +40,22 @@ public class BlogController {
 
     @PutMapping("/blog")
     @ResponseBody
-    public String updateBlog(@RequestParam Long id, @RequestBody BlogRequestDto requestDto) {
+    public String updateBlog(@RequestParam Long id, @RequestBody BlogRequestDto requestDto, @RequestHeader String tokenValue) {
+        String token = jwtUtil.substringToken(tokenValue);
+
+        if(!jwtUtil.validateToken(token)){
+            throw new IllegalArgumentException("Token Error");
+        }
         return blogService.updateBlog(id, requestDto);
     }
 
     @DeleteMapping("/blog")
-    public ResponseEntity<String> deleteBlog(@RequestParam Long id) {
+    public ResponseEntity<String> deleteBlog(@RequestParam Long id, @RequestHeader String tokenValue) {
+        String token = jwtUtil.substringToken(tokenValue);
+
+        if(!jwtUtil.validateToken(token)){
+            throw new IllegalArgumentException("Token Error");
+        }
         blogService.deleteBlog(id);
         return ResponseEntity.status(HttpStatus.OK).body("삭제 성공");
     }
